requirements.txt
selenium==4.27.1
Pillow==11.0.0
numpy==2.2.0
pandas==2.2.3
scikit-learn==1.5.2
matplotlib
plotly
seaborn
openpyxl
statsmodels
prophet
torch
kaleido
streamlit
streamlit-extras
pymysql